api/requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
django-rest-framework==0.1.0
django-storages==1.14.2
djangochannelsrestframework==1.1.0
djangochannels==4.0.0
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
